[Server] Implement interactivity and colors!
Supports hot-reloading of data, exiting, and showing a help screen. Want to add testing
devices later, just don't want to do it at the moment.

This requires a lot of terminal attributes to be set and unset, so ymmv when using a terminal
that doesn't support these.

Working on reverting to a most-terminal-friendly design when requesting no colors or
interactivity.